fixes and new README with final features and info

+++ packages/api/tests/setup.js
@@ -2,9 +2,15 @@ require("dotenv").config({ path: ".env.test" });
 const { sequelize } = require("../config/db");
 
 beforeAll(async () => {
+  console.log(`Running tests with DB: ${process.env.MYSQL_DATABASE}`);
+  if (!sequelize) {
+    throw new Error("Sequelize instance is undefined");
+  }
   await sequelize.sync({ force: true });
 });
 
 afterAll(async () => {
-  await sequelize.close();
+  if (sequelize) {
+    await sequelize.close();
+  }
 });
